+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* **Bugfix**: Ensure that *all* methods decorated with `@init()` in *all* classes in an inheritance chain only fire when the *last* enhanced constructor finishes.
 * **Other**: Bump depdencies, play some code golf, tweak error messages
 
 ## v 1.0.0